+++ internal/privileges/builtin.go
@@ -56,16 +56,16 @@ func init() {
 		Name:    "GLOBAL DEFAULT",
 		Scope:   "database",
 		Inspect: inspectGlobalDefault,
-		Grant:   `ALTER DEFAULT PRIVILEGES FOR ROLE <owner> GRANT <privilege> ON <acl> TO <grantee>;`,
-		Revoke:  `ALTER DEFAULT PRIVILEGES FOR ROLE <owner> REVOKE <privilege> ON <acl> FROM <grantee>;`,
+		Grant:   `ALTER DEFAULT PRIVILEGES FOR ROLE <owner> GRANT <privilege> ON <object> TO <grantee>;`,
+		Revoke:  `ALTER DEFAULT PRIVILEGES FOR ROLE <owner> REVOKE <privilege> ON <object> FROM <grantee>;`,
 	}.MustRegister()
 	ACL{
 		// implementation is chosed by name instead of scope.
 		Name:    "SCHEMA DEFAULT",
 		Scope:   "schema",
 		Inspect: inspectSchemaDefault,
-		Grant:   `ALTER DEFAULT PRIVILEGES FOR ROLE <owner> IN SCHEMA <schema> GRANT <privilege> ON <acl> TO <grantee>;`,
-		Revoke:  `ALTER DEFAULT PRIVILEGES FOR ROLE <owner> IN SCHEMA <schema> REVOKE <privilege> ON <acl> FROM <grantee>;`,
+		Grant:   `ALTER DEFAULT PRIVILEGES FOR ROLE <owner> IN SCHEMA <schema> GRANT <privilege> ON <object> TO <grantee>;`,
+		Revoke:  `ALTER DEFAULT PRIVILEGES FOR ROLE <owner> IN SCHEMA <schema> REVOKE <privilege> ON <object> FROM <grantee>;`,
 	}.MustRegister()
 
 	g := `GRANT <privilege> ON <acl> <schema> TO <grantee>;`
@@ -134,13 +134,13 @@ func registerRelationBuiltinProfile(class string, types ...string) {
 	for _, privType := range types {
 		TYPE := strings.ToUpper(privType)
 		BuiltinsProfiles["__default_"+privType+"_on_"+class+"__"] = []interface{}{map[string]interface{}{
-			"default": "global",
-			"type":    TYPE,
-			"on":      CLASS,
+			"type":   TYPE,
+			"on":     "GLOBAL DEFAULT",
+			"object": CLASS,
 		}, map[string]interface{}{
-			"default": "schema",
-			"type":    TYPE,
-			"on":      CLASS,
+			"type":   TYPE,
+			"on":     "SCHEMA DEFAULT",
+			"object": CLASS,
 		}}
 		BuiltinsProfiles["__"+privType+"_on_all_"+class+"__"] = []interface{}{map[string]interface{}{
 			"type": TYPE,